tests: subsys: bl_storage remove counters testing
This removes the testing of bl_monotonic_counters
in the bl_storage test. It creates a separate test
for it.

Ref: NCSDK-9045
