+++ tests/subsys/bootloader/bl_monotonic_counters/README
