+++ tests/subsys/bootloader/bl_monotonic_counters/prj.conf
@@ -1,0 +1,17 @@
+#
+# Copyright (c) 2020 Nordic Semiconductor ASA
+#
+# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+#
+
+CONFIG_ZTEST=y
+CONFIG_TEST_USERSPACE=n
+CONFIG_USERSPACE=n
+CONFIG_STDOUT_CONSOLE=n
+CONFIG_SECURE_BOOT=y
+CONFIG_SECURE_BOOT_STORAGE=y
+CONFIG_FW_INFO=y
+CONFIG_REBOOT=y
+CONFIG_FW_INFO_FIRMWARE_VERSION=10
+CONFIG_SECURE_BOOT_CRYPTO=y
+CONFIG_SB_NUM_VER_COUNTER_SLOTS=4
