+++ tests/subsys/bootloader/bl_monotonic_counters/testcase.yaml
@@ -1,0 +1,16 @@
+tests:
+  bootloader.bl_monotonic_counters:
+    platform_allow: nrf9160dk_nrf9160 nrf5340dk_nrf5340_cpuapp
+    integration_platforms:
+      - nrf9160dk_nrf9160
+      - nrf5340dk_nrf5340_cpuapp
+    tags: b0
+    harness: console
+    harness_config:
+      type: multi_line
+      ordered: true
+      regex:
+        - "Rebooting. Should fail to validate because of monotonic counter."
+        - "Firmware version \\(10\\) is smaller than monotonic counter \\(11\\)."
+        - "Failed to validate, permanently invalidating!"
+        - "No bootable image found\\. Aborting boot\\."
