+++ tests/subsys/bootloader/bl_storage/testcase.yaml
@@ -5,12 +5,3 @@ tests:
       - nrf9160dk_nrf9160
       - nrf5340dk_nrf5340_cpuapp
     tags: b0
-    harness: console
-    harness_config:
-      type: multi_line
-      ordered: true
-      regex:
-        - "Rebooting. Should fail to validate because of monotonic counter."
-        - "Firmware version \\(10\\) is smaller than monotonic counter \\(11\\)."
-        - "Failed to validate, permanently invalidating!"
-        - "No bootable image found\\. Aborting boot\\."
